+++ packages/selector-specificity/test/_import.mjs
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import selectorSpecificity from '@csstools/selector-specificity';
+import { selectorSpecificity } from '@csstools/selector-specificity';
 
 assert.equal(selectorSpecificity().a, 0);
 assert.equal(selectorSpecificity().b, 0);