+++ plugins/postcss-is-pseudo-class/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Is Pseudo Class
 
+### Unreleased (patch)
+
+- Update `@csstools/selector-specificity` (major)
+
 ### 2.0.4 (May 17, 2022)
 
  - Fix selector order with `:before` and other pseudo elements.